requirements.txt
torch==2.1.2
vllm==0.4.0
transformers==4.39.1
peft==0.14.0
tqdm==4.67.1
pandas==2.2.3
datasets==3.2.0
rouge==1.0.1
jieba==0.42.1
scikit-learn==1.6.1
autoawq==0.2.8